requirements.txt
# Dream Decoder Dependencies

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# NLP & ML
transformers>=4.36.0
torch>=2.1.0
spacy>=3.7.2

# Authentication
bcrypt>=4.1.2
PyJWT>=2.8.0

# Utilities
python-dateutil>=2.8.2
fpdf2>=2.8.5
langdetect>=1.0.9
googletrans==4.0.0-rc1
python-dotenv>=1.0.0
google-generativeai>=0.3.0

